Add Solr, replace React with Inferno, and update pip
This commit introduces Solr as the new full text search platform, which means
that the current implementation no longer relies on Postgres. As a consequence,
all the current Postgres DB only manages user auth (and later as a backup).
The prior tests for searching the DB were removed. ReactJS is now replaced with
InfernoJS, which will now represent an overall speedup in front-end operations.
Lastly, all pip packages and dockerize (for docker-compose) were updated.